refactor: move into files, add file descriptions for insights, and change props typing
